backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
groq
fastembed
python-multipart
pypdf
psycopg2-binary
pgvector
PyJWT
bcrypt
